target/rk35xx: disable overlayfs index
Fix docker migration.

By default, we use /overlay/upper/opt/docker as the docker data root.
If overlayfs index is enabled, docker will not be able to use overlay2 as the storage driver (generating the log "level=error msg="failed to mount overlay: device or resource busy" storage-driver=overlay2"), and will fall back to vfs.
After the docker data is migrated, overlay2 will definitely be used as the storage driver.
This will cause the storage driver to be inconsistent before and after the docker data migration, and all containers and images will disappear after the migration.